fix(plasma-new-hope): remove styled-components from Modal
