+++ packages/plasma-new-hope/src/components/Modal/Modal/Modal.styles.ts
@@ -1,0 +1,7 @@
+import { css } from '@linaria/core';
+
+export const noScroll = css`
+    :global(body) {
+        overflow-y: hidden;
+    }
+`;
