+++ packages/plasma-new-hope/src/components/Modal/Modal/variations/_view/base.ts
