+++ packages/plasma-new-hope/src/components/Modal/index.ts
@@ -1,2 +1,2 @@
-export { modalConfig, modalRoot } from './Modal';
-export { classes as modalClasses } from './Modal.tokens';
+export { modalConfig, modalRoot } from './Modal/Modal';
+export { classes as modalClasses } from './Modal/Modal.tokens';
